api/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.4
requests==2.31.0
python-dotenv==1.0.1
openai==1.68.2  # Compatible with langchain-openai
streamlit==1.35.0
langchain==0.3.24
langchain-core==0.3.55
langchain-openai==0.3.14
langsmith==0.3.21
